Déplacement des composant de la page Login, correction des imports et refactorisation
